+++ src/index.ts
@@ -1,12 +1,14 @@
-import "./utils/dotEnv";
-
-import * as profiler from "@google-cloud/profiler";
-import * as traceAgent from "@google-cloud/trace-agent";
-if (process.env.NODE_ENV === "production") {
+import * as traceAgent from '@google-cloud/trace-agent';
+if (process.env.NODE_ENV === 'production') {
   traceAgent.start();
+}
+import './utils/dotEnv';
+
+import * as profiler from '@google-cloud/profiler';
+if (process.env.NODE_ENV === 'production') {
   profiler.start();
 }
 
-import server from "./server";
+import server from './server';
 
 server();