+++ src/interfaces/data_sources.ts
@@ -1,0 +1,6 @@
+import { DataSource } from 'apollo-datasource';
+import LastFMAPI from 'services/lastfm/index';
+
+export interface DataSources extends DataSource {
+  lastfmAPI: LastFMAPI;
+}