+++ src/resolvers/index.ts
@@ -1,7 +1,9 @@
-import rootQuery from "./rootQuery";
+import rootQuery from './rootQuery';
+import user from './user';
 
 const allResolvers = {
   ...rootQuery,
+  ...user,
 };
 
 export default allResolvers;