+++ src/resolvers/rootQuery.ts
@@ -1,14 +1,63 @@
-import admin from "../services/firebase";
+import { DocumentSnapshot } from '@google-cloud/firestore';
+import { APIContext } from 'interfaces/api_context';
+import { FeatureArgs } from '../interfaces/feature_args';
 import { FeaturesArgs } from '../interfaces/features_args';
-import { QuerySnapshot, DocumentSnapshot } from "@google-cloud/firestore";
 import { Service } from '../interfaces/service';
 import { ServiceArgs } from '../interfaces/service_args';
-import { FeatureArgs } from '../interfaces/feature_args';
+import admin from '../utils/firebase';
 
 const rootQuery = {
   Query: {
+    async feature(_: void, args: FeatureArgs) {
+      const featureRef = await admin.firestore().collection('features').doc(args.project).get();
+      const data = featureRef.data();
+      const key: string = Object.keys(data).find((feature) => {
+        return feature === args.feature;
+      });
+      return {
+        enabled: data[key],
+        name: key,
+        project: args.project,
+      };
+    },
+    async features(_: void, args: FeaturesArgs, context: APIContext) {
+      const featureProjectsRef = await admin.firestore().collection('features');
+      // tslint:disable-next-line: max-line-length
+      const projects = (args.project) ? await featureProjectsRef.doc(args.project).get() : await featureProjectsRef.get();
+      if (projects instanceof DocumentSnapshot) {
+        context.logger.info(`Getting feature switches for ${projects.id}`);
+        const data = projects.data();
+        const keys = Object.keys(data);
+        const features = keys.map((key) => {
+          return {
+            enabled: data[key],
+            name: key,
+            project: projects.id,
+          };
+        });
+        return features;
+      }
+      const featureArray = projects.docs.flatMap((project, index) => {
+        context.logger.info(`Getting feature switches for ${project.id}`);
+        const data = project.data();
+        const keys = Object.keys(data);
+        const features = keys.map((key) => {
+          return {
+            enabled: data[key],
+            name: key,
+            project: project.id,
+          };
+        });
+        return features;
+      }) as [];
+
+      return featureArray;
+    },
+    user(_: void) {
+      return '';
+    },
     async services(_: void, args: ServiceArgs) {
-      const services = await admin.firestore().collection("services").get();
+      const services = await admin.firestore().collection('services').get();
       const serviceArray = services.docs.map((service) => service.data()) as [];
 
       return Object.keys(args).length === 0 ? serviceArray : serviceArray.filter((service: Service) => {
@@ -20,50 +69,6 @@ const rootQuery = {
         return true;
       });
     },
-    async features(_: void, args: FeaturesArgs) {
-      const featureProjectsRef = await admin.firestore().collection("features");
-      const projects = (args.project) ? await featureProjectsRef.doc(args.project).get() : await featureProjectsRef.get();
-      if (projects instanceof DocumentSnapshot) {
-        console.info(`Getting feature switches for ${projects.id}`);
-        const data = projects.data();
-        const keys = Object.keys(data);
-        const features = keys.map(key => {
-          return {
-            project: projects.id,
-            name: key,
-            enabled: data[key],
-          }
-        })
-        return features;
-      }
-      const featureArray = projects.docs.flatMap((project, index) => {
-        console.info(`Getting feature switches for ${project.id}`);
-        const data = project.data()
-        const keys = Object.keys(data);
-        const features = keys.map(key => {
-          return {
-            project: project.id,
-            name: key,
-            enabled: data[key],
-          }
-        })
-        return features;
-      }) as [];
-
-      return featureArray;
-    },
-    async feature(_: void, args: FeatureArgs) {
-      const featureRef = await admin.firestore().collection("features").doc(args.project).get();
-      const data = featureRef.data();
-      const key: string = Object.keys(data).find((feature) => {
-        return feature === args.feature;
-      });
-      return {
-        project: args.project,
-        name: key,
-        enabled: data[key]
-      }
-    }
   },
 };