+++ src/resolvers/user/index.ts
@@ -1,0 +1,9 @@
+import music from './music';
+import userResolver from './User';
+
+const user = {
+  ...userResolver,
+  ...music,
+};
+
+export default user;