+++ src/resolvers/user/music/index.ts
@@ -1,0 +1,9 @@
+import musicResolver from './Music';
+import track from './Track';
+
+const music = {
+  ...musicResolver,
+  ...track,
+};
+
+export default music;