+++ src/schema/services/service.ts
@@ -1,10 +1,17 @@
 const Service = `
+  "service type"
   type Service {
+    "Name of the service"
     name: String
+    "Language the service is written in e.g. Ruby"
     language: String
+    "URL of the service"
     url: String
+    "Host of the service"
     host: String
+    "Description of the service"
     description: String
+    "Status of the service"
     status: String
   }
 `;