+++ src/schema/user/Music/index.ts
@@ -1,0 +1,9 @@
+import music from './Music';
+import track from './Track';
+
+const Music = `
+  ${music}
+  ${track}
+`;
+
+export default Music;