+++ src/schema/user/User.ts
@@ -1,0 +1,9 @@
+const user = `
+  "User type - describes the user"
+  type User {
+    "Music related information"
+    music: Music
+  }
+`;
+
+export default user;