+++ src/server/app.ts
@@ -1,24 +1,24 @@
-import Sentry from "@sentry/node";
-import cors from "cors";
-import express from "express";
-import helmet from "helmet";
+import * as Sentry from '@sentry/node';
+import cors from 'cors';
+import express from 'express';
+import helmet from 'helmet';
 
 const app = express();
-app.enable("trust proxy");
+app.enable('trust proxy');
 
-const defaultOrigin: string[] = process.env.NODE_ENV !== "production" ? ["https://localhost"] : [];
-const origin = process.env.CORS_URLS ? process.env.CORS_URLS.split(",") : ["https://mattriley.info"];
+const defaultOrigin: string[] = process.env.NODE_ENV !== 'production' ? ['https://localhost'] : [];
+const origin = process.env.CORS_URLS ? process.env.CORS_URLS.split(',') : ['https://mattriley.info'];
 
 const corsOptions = {
-  allowedHeaders: ["Authorization", "Content-Type"],
-  methods: ["GET", "POST"],
+  allowedHeaders: ['Authorization', 'Content-Type'],
+  methods: ['GET', 'POST'],
   origin: defaultOrigin.concat(origin),
   preflightContinue: false,
 };
 
 app.use(cors(corsOptions));
 
-if (process.env.NODE_ENV === "production") {
+if (process.env.NODE_ENV === 'production') {
   app.use(helmet.hsts({
     maxAge: 31536000,
     preload: true,
@@ -28,8 +28,8 @@ if (process.env.NODE_ENV === "production") {
   app.use(Sentry.Handlers.errorHandler() as express.ErrorRequestHandler);
 }
 
-app.disable("x-powered-by");
+app.disable('x-powered-by');
 
-app.set("port", process.env.PORT || 3000);
+app.set('port', process.env.PORT || 3000);
 
 export default app;