+++ src/server/index.ts
@@ -1,13 +1,13 @@
-import http from "http";
+import http from 'http';
 
-import apollo from "./apollo";
-import app from "./app";
+import logger from '../utils/logger';
+import apollo from './apollo';
+import app from './app';
 
-apollo.applyMiddleware({ app, path: "/" });
+apollo.applyMiddleware({ app, path: '/' });
 
-const server = () => http.createServer(app).listen(app.get("port"), () => {
-  // tslint:disable-next-line:no-console
-  console.info("Express server listening on port " + app.get("port"));
+const server = () => http.createServer(app).listen(app.get('port'), () => {
+  logger.info('Express server listening on port ' + app.get('port'));
 });
 
 export default server;