+++ src/utils/firebase.ts
@@ -1,4 +1,4 @@
-import admin from "firebase-admin";
+import admin from 'firebase-admin';
 
 admin.initializeApp({
   credential: admin.credential.applicationDefault(),