+++ src/utils/logger.ts
@@ -1,0 +1,16 @@
+import { LoggingWinston } from '@google-cloud/logging-winston';
+import { createLogger, transports } from 'winston';
+
+const loggingWinston = new LoggingWinston();
+
+const logger = createLogger({
+  level: 'info',
+  transports: [
+    new transports.Console({
+      handleExceptions: true,
+    }),
+    loggingWinston,
+  ],
+});
+
+export default logger;