+++ CHANGELOG.md
@@ -1,6 +1,22 @@
-# Change Log
+# Changelog
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
+
+## [1.1.0](https://github.com/matt-riley/api.mattriley.info/compare/v1.0.0...v1.1.0) (2019-06-02)
+
+
+### Bug Fixes
+
+* **Fixed errors:** Would not work on App Engine, resolved the issues ([4f8a1a8](https://github.com/matt-riley/api.mattriley.info/commit/4f8a1a8))
+
+
+### Features
+
+* **Caching enabled:** The allows persisted queries, caching of responses and fields in memcache on ([5012ed4](https://github.com/matt-riley/api.mattriley.info/commit/5012ed4))
+* **Logger:** Adds winston as the logger and adds it to the context ([90a5d09](https://github.com/matt-riley/api.mattriley.info/commit/90a5d09))
+* **Recently listened tracks:** Addes recently listened tracks from Last.fm ([f2aaf51](https://github.com/matt-riley/api.mattriley.info/commit/f2aaf51))
+
+
 
 # [1.0.0](https://github.com/matt-riley/mattapps/tree/master/api.mattriley.info/compare/v0.0.1...v1.0.0) (2019-04-27)